pizza info sql update

+++ src/main/java/groupId/artifactId/dao/entity/api/IPizzaInfo.java
@@ -5,5 +5,5 @@ public interface IPizzaInfo {
 
     String getDescription();
 
-    int getSize();
+    Long getSize();
 }
